Remove old way generation, remove legacy and unused code
