+++ invokeai/app/util/step_callback.py
@@ -1,10 +1,38 @@
+import torch
+from PIL import Image
 from invokeai.app.models.exceptions import CanceledException
 from invokeai.app.models.image import ProgressImage
 from ..invocations.baseinvocation import InvocationContext
 from ...backend.util.util import image_to_dataURL
-from ...backend.generator.base import Generator
 from ...backend.stable_diffusion import PipelineIntermediateState
 
+
+def sample_to_lowres_estimated_image(samples):
+    # origingally adapted from code by @erucipe and @keturn here:
+    # https://discuss.huggingface.co/t/decoding-latents-to-rgb-without-upscaling/23204/7
+
+    # these updated numbers for v1.5 are from @torridgristle
+    v1_5_latent_rgb_factors = torch.tensor(
+        [
+            #    R        G        B
+            [0.3444, 0.1385, 0.0670],  # L1
+            [0.1247, 0.4027, 0.1494],  # L2
+            [-0.3192, 0.2513, 0.2103],  # L3
+            [-0.1307, -0.1874, -0.7445],  # L4
+        ],
+        dtype=samples.dtype,
+        device=samples.device,
+    )
+
+    latent_image = samples[0].permute(1, 2, 0) @ v1_5_latent_rgb_factors
+    latents_ubyte = (
+        ((latent_image + 1) / 2)
+        .clamp(0, 1)  # change scale from -1..1 to 0..1
+        .mul(0xFF)  # to 0..255
+        .byte()
+    ).cpu()
+
+    return Image.fromarray(latents_ubyte.numpy())
 
 def stable_diffusion_step_callback(
     context: InvocationContext,
@@ -37,7 +65,7 @@ def stable_diffusion_step_callback(
     #     step = intermediate_state.step
 
     # TODO: only output a preview image when requested
-    image = Generator.sample_to_lowres_estimated_image(sample)
+    image = sample_to_lowres_estimated_image(sample)
 
     (width, height) = image.size
     width *= 8
