--- invokeai/backend/generator/__init__.py
@@ -1,13 +1,0 @@
-"""
-Initialization file for the invokeai.generator package
-"""
-from .base import (
-    InvokeAIGenerator,
-    InvokeAIGeneratorBasicParams,
-    InvokeAIGeneratorOutput,
-    Txt2Img,
-    Img2Img,
-    Inpaint,
-    Generator,
-)
-from .inpaint import infill_methods
