--- invokeai/backend/web/__init__.py
@@ -1,4 +1,0 @@
-"""
-Initialization file for the web backend.
-"""
-from .invoke_ai_web_server import InvokeAIWebServer
